+++ lib/dvla/browser/drivers.rb
@@ -1,3 +1,7 @@
+require_relative 'drivers/builders/builder'
+require_relative 'drivers/builders/cuprite_builder'
+require_relative 'drivers/builders/selenium_builder'
+
 require_relative 'drivers/mobile_profiles'
 require_relative 'drivers/meta_drivers'
 require_relative 'drivers/version'
@@ -7,5 +11,24 @@ require 'capybara/cuprite'
 require 'selenium-webdriver'
 
 module DVLA
-  class Error < StandardError; end
+  module Browser
+    module Drivers
+      def self.cuprite_builder(config = nil)
+        CupriteBuilder.new(config)
+      end
+
+      def self.selenium_builder(config = nil)
+        SeleniumBuilder.new(config)
+      end
+
+      def self.builder(driver, config: nil)
+        case driver.to_s.downcase.to_sym
+        when :cuprite
+          CupriteBuilder.new(config)
+        else
+          SeleniumBuilder.new(config)
+        end
+      end
+    end
+  end
 end