+++ lib/dvla/browser/drivers/tasks.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
 load File.expand_path('tasks/browser.rake', __dir__)
+load File.expand_path('tasks/browser_builder.rake', __dir__)